requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
cffi==1.14.1
chardet==3.0.4
cryptography==3.0
Django==3.0.6
idna==2.9
mysqlclient==1.4.6
pycparser==2.20
PyMySQL==0.9.3
pytz==2020.1
requests==2.23.0
six==1.15.0
sqlparse==0.3.1
ujson==3.0.0
urllib3==1.25.9
atlassian-python-api==1.16.1
